breaking: rename histogram(s->'').py
