+++ pymatviz/histogram.py
