+++ tests/test_histogram.py
